backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
